third_party/blink/web_tests/fast/css/shorthands-four-values-expected.txt

Test for WebKit bug 7987: Inspector displays incorrect summary for padding information

PASS test0.style.padding is "20px"
PASS test1.style.padding is "20px 30px"
PASS test2.style.padding is "20px 30px 10px 15px"
PASS test3.style.padding is "20px 30px"
PASS test4.style.borderColor is "yellow red red"
PASS test5.style.borderColor is "green blue red"
PASS test6.style.borderColor is "blue yellow red green"
PASS test7.style.borderColor is "red blue green yellow"
PASS test8.style.borderWidth is "2px 5px 2px 2px"
PASS test9.style.borderWidth is "2px 4px 2px 3px"
PASS test10.style.borderWidth is "2px 3px 5px"
PASS test11.style.borderWidth is "20px 30px 55px 60px"
PASS test12.style.borderStyle is "solid solid dashed"
PASS test13.style.borderStyle is "solid dashed dotted"
PASS test14.style.borderStyle is "dotted dashed"
PASS test15.style.borderStyle is ""
PASS test16.style.margin is "20px 20px 20px 25px"
PASS test17.style.margin is "20px 30px 20px 50px"
PASS test18.style.margin is "20px 30px 55px"
PASS test19.style.margin is ""
PASS successfullyParsed is true

TEST COMPLETE
